Support multiple validation loaders
The motivation for this (also implemented here) is to check whether
changing a different pixel increases consistency loss, as an ablation.
Turns out it doesn't, so it seems the backdoor detection works!

One weird thing is that loss with a different corner pixel set is
in fact *lower* than on the training distribution, where no pixels
are changed. That's pretty weird, so worth digging into.