Add abstraction eval script
This will let us quickly run new evaluations without retraining,
or conversely retrain without running a large suite of expensive
evaluations.

To make this possible, I added a bunch of better logging and
model saving. The philosophy is that every training run
has some directory where it stores the final model weights
and all hparams. The evaluation script should only need to know
that directory and then figure out the rest. The eval script
will store its metrics in a file in the training directory,
instead of creating its own run directory, so that everything
about a training run is in one place.