Refactor data loading to support Gaussian noise
I wanted to check whether an input with noise would be flagged
as anomalous (it isn't for reasonable amounts of noise). To do
that, I refactored a lot of the data transforms, should be more
flexible and robust now. In particular, inputs will always have
an info dict, which is added first by its own transform.