Add simple finetuning detector
I upgraded the minimum Jax version to 0.4.16 because the new code
needs the KL divergence in jax.scipy.special. Also had to make some
small changes because the finetuning detector doesn't support layerwise
scores, which some of the eval code assumed. The detector can probably
be improved (e.g. it's unclear whether the current KL divergence is
a good metric, and I didn't optimize the hyperparameters at all).
But it's already a decent baseline, e.g. it get ~0.8 AUC for a
WaNet/CIFAR10 model for which the Mahalanobis baseline appears to be worse
than random (?!)