Mostly circumvent MPS memory leak
We're now just padding all inputs to the maximum sequence length as a workaround. Lots of room for improvement here, e.g. would be great if TransformerLens let us pass kwargs through to to_tokens (to save a bunch of copied code), and ideally we'd use the maximum sequence length over all training examples instead. Also looks like memory usage is still increasing slightly, but it's now on the order of 200MB for one epoch on the IMDB dataset with Pythia-14m, so at least some local testing is totally feasible now. For serious training, we'll use CUDA anyway (but this remaining memory leak might of course also affect that if it's a different source, should keep an eye out)